backend/: split sfx package into client.go and request.go
