+++ backend/sfx/client_test.go
@@ -1,0 +1,88 @@
+package sfx
+
+import (
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"reflect"
+	"testing"
+)
+
+func TestIsValidXML(t *testing.T) {
+	var tests = []struct {
+		testXMLFile string
+		expected    bool
+	}{
+		{"./testdata/sfx-context-object-valid.xml", true},
+		{"./testdata/sfx-context-object-invalid-truncated.xml", false},
+		{"", false},
+	}
+
+	for _, tt := range tests {
+		testname := fmt.Sprintf("%s", tt.testXMLFile)
+		t.Run(testname, func(t *testing.T) {
+			data, _ := ioutil.ReadFile(tt.testXMLFile)
+			ans := isValidXML(data)
+			if ans != tt.expected {
+				t.Errorf("isValidXML returned '%v', expecting '%v'", ans, tt.expected)
+			}
+		})
+	}
+}
+
+func TestParseOpenURL(t *testing.T) {
+	var tests = []struct {
+		queryString map[string][]string
+		expected    map[string][]string
+		expectedErr error
+	}{
+		{map[string][]string{"genre": {"book"}, "rft.genre": {"book"}}, openURL{"genre": {"book"}}, nil},
+		{map[string][]string{"genre": {"book"}, "rft.genre": {"journal", "book"}}, openURL{"genre": {"journal", "book"}}, nil},
+		{map[string][]string{"genre": {"book"}, "rft.genre": {"journal"}}, openURL{"genre": {"journal"}}, nil},
+		{map[string][]string{"genre": {"book"}}, openURL{}, errors.New("error")},
+	}
+
+	for _, tt := range tests {
+		testname := fmt.Sprintf("%s", tt.queryString)
+
+		t.Run(testname, func(t *testing.T) {
+			ans, err := parseOpenURL(tt.queryString)
+			if reflect.DeepEqual(ans, tt.expected) {
+				t.Errorf("parseOpenURL returned '%v', expecting '%v'", ans, tt.expected)
+			}
+			if tt.expectedErr != nil {
+				if err == nil {
+					t.Errorf("parseOpenURL err was '%v', expecting '%v'", err, tt.expectedErr)
+				}
+			}
+		})
+	}
+}
+
+func TestValidGenre(t *testing.T) {
+	var tests = []struct {
+		genre       []string
+		expected    string
+		expectedErr error
+	}{
+		{[]string{"book", "book"}, "book", nil},
+		{[]string{"journal", "book"}, "journal", nil},
+		{[]string{"book"}, "book", nil},
+		{[]string{"wrong"}, "", errors.New("error")},
+	}
+
+	for _, tt := range tests {
+		testname := fmt.Sprintf("%s", tt.genre)
+		t.Run(testname, func(t *testing.T) {
+			ans, err := validGenre(tt.genre)
+			if ans != tt.expected {
+				t.Errorf("validGenre returned '%v', expecting '%v'", ans, tt.expected)
+			}
+			if tt.expectedErr != nil {
+				if err == nil {
+					t.Errorf("validGenre err was '%v', expecting '%v'", err, tt.expectedErr)
+				}
+			}
+		})
+	}
+}
